--- packages/ui/src/views/user-management/context/dialogs/index.ts
@@ -1,1 +1,0 @@
-export * from './context'
