+++ packages/ui/src/views/user-management/providers/DialogsProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './DialogsProvider'
+export * from './hooks/useDialogs'
