+++ packages/ui/src/views/user-management/user-management-page.tsx
@@ -2,11 +2,11 @@ import { SandboxLayout } from '@/views'
 import { Dialogs } from '@/views/user-management/components/dialogs'
 import { Content } from '@/views/user-management/components/page-components/content'
 import { UserManagementTabs } from '@/views/user-management/components/tabs'
-import { DialogsProvider } from '@/views/user-management/context/dialogs/context'
 import { useUserManagement } from '@/views/user-management/hooks/use-user-management'
 import { IUserManagementPageProps } from '@/views/user-management/types'
 import { getFilterOptions, getSortDirections, getSortOptions } from '@views/repo/constants/filter-options'
 import { useFilters } from '@views/repo/hooks'
+import { DialogsProvider } from '@views/user-management/providers/DialogsProvider'
 
 export const UserManagementPage: React.FC<IUserManagementPageProps> = ({
   useAdminListUsersStore,
